small improvements

+++ src/app/api/counterparty/delete.ts
@@ -5,14 +5,14 @@ import { Pool } from "pg";
 export async function deleteCounterparty(
   counterpartyId: number,
   userId: number
-) {
+): Promise<{ success: boolean }> {
   const client: Pool = new Pool();
-  await client.query(`DELETE FROM public.counterparty WHERE id = $1;`, [
-    counterpartyId,
+  const res = await client.query(`DELETE FROM public.counterparty WHERE id = $1 & user_id = $2;`, [
+    counterpartyId, userId
   ]);
   await client.end();
 
   return {
-    createTransaction: true,
+    success: true,
   };
 }
